requirements.txt
tiktoken~=0.5.2
openai~=1.5.0
python-dotenv~=1.0.0
langchain~=0.0.351